Adapted citation format to those of similar origin
We have OASiS citation prefs for IETF RFCs and ISO standards as well, also W3C  suggest a format. This leads to some entropy anyway.
One thing we can maintain is the ordering of name parts given first family second, and that with softer refs, the title starts the entry.