Clarify the association of a product to a score
A CVSS score listed for a product should reflect that product's status.
For example, a fixed product version should in most cases have a CVSS score
of 0 since the vulnerability in it no longer exists. Previous version
ranges of that same product can list the CVSS scores as it applies to them.